Add routing to policy sign
