+++ lib/extension/bridge/context.ts
@@ -1,3 +1,4 @@
 export type BridgeContext = {
+  requestFavicon?: string
   requestOrigin: string
 }
